refactor: extract BlogIndexPage.tsx
