+++ src/components/routes/blog/index.ts
@@ -1,1 +1,2 @@
 export { BlogPostPage } from "./BlogPostPage";
+export { BlogIndexPage } from "./BlogIndexPage";
